update the cassandra config files at docker entrypoint

+++ catalog/cassandra/3.11/dockerfile/docker-entrypoint.sh
@@ -4,21 +4,23 @@ set -e
 DATA_DIR=/data
 COMMITLOG_DIR=/journal
 CONFIG_DIR=$DATA_DIR/conf
+
+# after release 0.9.5, cassandra.yaml, rackdc.properties and jvm.options are not created any more.
+# instead, we will read the configs from service.conf and sys.conf, and update the default files.
 CASSANDRA_YAML_FILE=$CONFIG_DIR/cassandra.yaml
 CASSANDRA_RACKDC_FILE=$CONFIG_DIR/cassandra-rackdc.properties
-CASSANDRA_LOG_FILE=$CONFIG_DIR/logback.xml
 CASSANDRA_JVM_FILE=$CONFIG_DIR/jvm.options
+
+CASSANDRA_LOG_FILE=$CONFIG_DIR/logback.xml
 CASSANDRA_JMXREMOTEPASSWD_FILE=$CONFIG_DIR/jmxremote.password
 syscfgfile=$CONFIG_DIR/sys.conf
+servicecfgfile=$CONFIG_DIR/service.conf
 
-# sanity check to make sure the volume is mounted to /data.
+# Sanity check to make sure the volume is mounted to /data and /journal.
+# If volume is not mounted for any reason, exit. This ensures data will not be written
+# to the local directory, and get lost when the container moves to another node.
 if [ ! -d "$DATA_DIR" ]; then
   echo "error: $DATA_DIR not exist. Please make sure the volume is mounted to $DATA_DIR." >&2
-  exit 1
-fi
-# sanity check to make sure the db config files are created.
-if [ ! -d "$CONFIG_DIR" ]; then
-  echo "error: $CONFIG_DIR not exist." >&2
   exit 1
 fi
 if [ ! -d "$COMMITLOG_DIR" ]; then
@@ -26,12 +28,9 @@ if [ ! -d "$COMMITLOG_DIR" ]; then
   exit 1
 fi
 
-if [ ! -f "$CASSANDRA_YAML_FILE" ]; then
-  echo "error: $CASSANDRA_YAML_FILE not exist." >&2
-  exit 1
-fi
-if [ ! -f "$CASSANDRA_RACKDC_FILE" ]; then
-  echo "error: $CASSANDRA_RACKDC_FILE not exist." >&2
+# sanity check to make sure the db config files are created.
+if [ ! -d "$CONFIG_DIR" ]; then
+  echo "error: $CONFIG_DIR not exist." >&2
   exit 1
 fi
 if [ ! -f "$CASSANDRA_LOG_FILE" ]; then
@@ -46,20 +45,63 @@ if [ ! -f "$syscfgfile" ]; then
 fi
 
 if [ "$(id -u)" = '0' ]; then
-  cp $CASSANDRA_YAML_FILE /etc/cassandra/
-  cp $CASSANDRA_RACKDC_FILE /etc/cassandra/
-  cp $CASSANDRA_LOG_FILE /etc/cassandra/
+  if [ -f "$CASSANDRA_YAML_FILE" ]; then
+    cp $CASSANDRA_YAML_FILE /etc/cassandra/
+    cp $CASSANDRA_RACKDC_FILE /etc/cassandra/
+  fi
   # jvm.options file does not exist before release 0.9.1
   if [ -f "$CASSANDRA_JVM_FILE" ]; then
     cp $CASSANDRA_JVM_FILE /etc/cassandra/
   fi
+
+  cp $CASSANDRA_LOG_FILE /etc/cassandra/
   # jmxremote.password file does not exist before release 0.9.2
   if [ -f "$CASSANDRA_JMXREMOTEPASSWD_FILE" ]; then
     export LOCAL_JMX="no"
     cp $CASSANDRA_JMXREMOTEPASSWD_FILE /etc/cassandra/
   fi
-fi
 
+  # after 0.9.5, replace the fields in the default cassandra config files.
+  if [ -f "$servicecfgfile" ]; then
+    # source the sys and service config files
+    . $syscfgfile
+    . $servicecfgfile
+
+    # replace the configs in cassandra.yaml
+    DefaultYaml="/etc/cassandra/cassandra.yaml"
+    sed -ri 's/(cluster_name:).*/\1 '\'$CLUSTER\''/' $DefaultYaml
+
+    sed -i 's/# hints_directory:.*/hints_directory: \/data\/hints/g' $DefaultYaml
+    sed -i 's/- \/var\/lib\/cassandra\/data/- \/data\/data/g' $DefaultYaml
+    sed -i 's/commitlog_directory:.*/commitlog_directory: \/journal/g' $DefaultYaml
+    sed -i 's/saved_caches_directory:.*/saved_caches_directory: \/data\/saved_caches/g' $DefaultYaml
+
+    sed -ri 's/(- seeds:).*/\1 "'"$CASSANDRA_SEEDS"'"/' $DefaultYaml
+
+    ListenAddr=$SERVICE_MEMBER
+    if [ "$PLATFORM" = "swarm" ]; then
+      ListenAddr=""
+    fi
+    sed -i 's/listen_address:.*/listen_address: '$ListenAddr'/g' $DefaultYaml
+    sed -i 's/broadcast_address:.*/broadcast_address: '$SERVICE_MEMBER'/g' $DefaultYaml
+    RpcAddr=$SERVICE_MEMBER
+    if [ "$PLATFORM" = "swarm" ]; then
+      RpcAddr="0.0.0.0"
+    fi
+    sed -i 's/rpc_address:.*/rpc_address: '$RpcAddr'/g' $DefaultYaml
+    sed -i 's/broadcast_rpc_address:.*/broadcast_rpc_address: '$SERVICE_MEMBER'/g' $DefaultYaml
+
+    sed -i 's/endpoint_snitch:.*/endpoint_snitch: GossipingPropertyFileSnitch/g' $DefaultYaml
+
+    sed -i 's/authenticator:.*/authenticator: PasswordAuthenticator/g' $DefaultYaml
+    sed -i 's/authorizer:.*/authorizer: CassandraAuthorizer/g' $DefaultYaml
+
+    # replace the configs in cassandra-rackdc.properties
+    RackdcFile="/etc/cassandra/cassandra-rackdc.properties"
+    sed -i 's/dc=.*/dc='$REGION'/g' $RackdcFile
+    sed -i 's/rack=.*/rack='$AVAILABILITY_ZONE'/g' $RackdcFile
+  fi
+fi
 
 # allow the container to be started with `--user`
 if [ "$1" = 'cassandra' -a "$(id -u)" = '0' ]; then
@@ -77,10 +119,15 @@ if [ "$1" = 'cassandra' -a "$(id -u)" = '0' ]; then
 	exec gosu cassandra "$BASH_SOURCE" "$@"
 fi
 
-# source the sys config file
+# source the sys and service config files
 . $syscfgfile
+if [ -f "$servicecfgfile" ]; then
+  . $servicecfgfile
+fi
 echo $SERVICE_MEMBER
+echo "$JVM_OPTS"
 echo
+
 
 # enable jolokia agent
 # TODO enable the basic auth, https://jolokia.org/reference/html/agents.html
